Moving toward unit-space everything on screen.
To get toward a world where I can resize the output video, I need to be able to resize the UI. The first step toward that is to make the UI pixel-size-agnostic. This means taking everything to 0.0-1.0 space. For the moment, I'm hard-coding positions of each widget as I convert them from taking pixel-space to unit-space arguments. As I make each change, I have to fix the widget itself to render itself and its text correctly. It's late, time for bed, time to commit today's work.

When I'm done with this part of it, I'll be making an arrangement manager of some sort that allows me to move widgets around. Switching between output sizes will either load the previous layout for that size or convert the current layout to that pixel size.