Switch screen space to [-1..1]
There's still some fiddling to do with some of the font users, who are still trying to place strings in their old (pixel space) locations. I'll also have to make some sense of how to valign text in things like the datetime widget.  (The font renderer still "thinks" in screen space coordinates. I need to decide if this is what I want, long-term. I borrowed the thing from elsewhere, and it thinks in terms of point (pixel) sizes.)

I'm still missing the hang glider glyph in the map... because I haven't done a thing to transition it between spaces. I suspect that'll be painful.