Get Widget background masks and selection borders working as intended.
Rendering of the background mask had been left to the default shader attached to each Widget. I'd rather have one way to render a mask and one way to render a border, and have those handled outside of the derived Widget types.

The current state has everything rendering the border. I need to create an actual widget manager that lets the user move the things around. I'm not looking forward to it. I may ignore it for a very long time, but at least I have the "selected widget" effect in place.